update (recent-component): add recent component to user page and refactor imports
