+++ client/src/routes/Index.tsx
@@ -13,11 +13,11 @@ export default function Index() {
                 <div className="s3 center-align">
                     <FiltersPannel posts={posts as Post[]} setPosts={setPosts} />
                 </div>
-                <div className="s6 center align">
+                <div className="s6 center-align">
                     <Feed posts={posts as Post[]} />
                 </div>
-                <div className="s3 center align">
-                    <RecentPannel/>
+                <div className="s3 center-align">
+                    <RecentPannel posts={posts as Post[]}/>
                 </div>
             </div>
         </>
