+++ client/src/routes/PostPage.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { useLoaderData } from "react-router-dom";
 import { BackButton } from "../components/BackButton";
 import { PostData } from "../components/PostData";
+import { Post } from "../types";
 
-export const Post: React.FC = () => {
+export const PostPage: React.FC = () => {
     const post = useLoaderData()
-    console.log(post)
 
     return (
         <>
@@ -14,7 +14,7 @@ export const Post: React.FC = () => {
                     <BackButton />
                 </div>
                 <div className="s6 center-align">
-                    <PostData post={post} />
+                    <PostData post={post as Post} />
                 </div>
                 <div className="s3"></div>
             </div>
